CRAS: HFP: Handle unexpected SLC disconnection
Whe the service level connection is established for HFP
device, there are some cases the read() call to the rfcomm
socket fails with errno ECONNRESET.
This happens sometime when the HFP headset powers off and
reproducible 100% of the time when the qualification test
terminates manually in TPS software.
However bluetooth daemon doesn't notify CRAS for this error
to release the disconnection. Handle this error by adding a
callback to destroy all HFP resources.

BUG=chromium:211636
TEST=Unittest. Run PTS qualification test on a DUT with flag
enable_hfp, manually terminates it during the middle of the
test and verify the HFP iodev disappear from UI tray.

Change-Id: I9d664a45f68a8f4dc92150e4f647f20ca6f46c5b
Reviewed-on: https://chromium-review.googlesource.com/194888
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>