UCM config: nyan - Increase input gain.
A previous commit dut back the input gain.  It turns out that the
ADC[L|R] registers affect both analog and digital input.  This leaves
the default input gain too low, below the range the AGC in hangouts
can fix, for the DMIC.

Put back 18dB of gain on the input path.

BUG=chrome-os-partner:29047
TEST=record and play back voice through both internal mic and headset.

Change-Id: Ic84c12834766d043e4bd3f023efc74de6967a3e1
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/202065
Reviewed-by: Owen Lin <owenlin@chromium.org>