CRAS: audio_thread: close devices when deactivating.
The device was in draining mode when we removed the last stream.
We should close it when it get removed from the active list.
Otherwise, it will keep playing the content in the hardware buffer.

The new active device doesn't work because the devs_open is still true,
so we didn't open it at all.

BUG=chromium:383026
TEST=run youtube on chromium os device

Change-Id: I69bda3a0672fe1622998b04c3f8f7105b6d0191f
Reviewed-on: https://chromium-review.googlesource.com/203361
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Owen Lin <owenlin@chromium.org>
Tested-by: Owen Lin <owenlin@chromium.org>