CRAS: audio thread: Limit thread buffer by the buffer size of dev
This change adds back the check to limit thread buffer not greater
then the buffer size of device, which was removed in
"df69e201 CRAS: iodev: Move cb_threshold, used_size to thread".
A regression was caused when watching Youtube with a2dp device,
in which case the size of device buffer(896) is less than the callback
threshold of client stream(1024), so that audio thread fetches audio
too fast and user will see video played almost a second ahead of audio
heard from a2dp headset.

BUG=chromium:386975
TEST=Watch YouTube with a2dp headset.

Change-Id: I4dfb49001fc80d5e27b79f1136382373e9643ba2
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/204828