CRAS: audio_thread - Base request size on available space.
Don't limit the number of frames to request based on the thread's
buffer size.  This is configured based on the lowest latency stream
and can cause trouble for high latency streams playing concurrently
with lower latency streams.  This will need modification for multiple
outputs, where each dev will have to be checked for available space.

BUG=none
TEST=play audio to USB and A2DP, check for dropouts.

Change-Id: I90029632c3bbaad7a9a087a4ba675a76f37a8ab0
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/206450