CRAS: audio_thread - drain based on device buffer size.
The buffer size of the device and the thread may be different, we want
to drain based on the size of the device's buffer.  If the device
buffer is smaller than the thread's, there is no way to fill it all
the way up and this could end in an infinite loop.

BUG=none
TEST=remove last stream while playing to bluetooth device, notice cras
doesn't eat an entire cpu forever.

Change-Id: I62bf3d36b712346764108d7d03f45b40f3ae8e03
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/206840