CRAS: HFP SLC: move storing number to telephony
To pass HFP qualification we need a test telephony module to
fulfill test specification and it should run independently
from HFP device and SLC.
This change decouple SLC and the telephony module so that some
actions can be tested. For example, store the number to dial
before a SLC is established.

BUG=chromium:211636
TEST=TC_AG_OCL_VC_01_I, TC_AG_OCL_VC_02_I

Change-Id: I1d8902adc2d08caf88ec4afeb8d7fb80bbb0e6be
Reviewed-on: https://chromium-review.googlesource.com/206987
Tested-by: MengHuan Yu <menghuan@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>