CRAS: HFP SLC: Handle call hold action
It is required to mock the 'call hold' action and correctly pass
up the status to pass some HFP qualification tests.
Add dbus method to expose this ability,
so when PTS software asks DUT to place a active call on hold,
execute command dbus-send for that.

BUG=chromium:211636
TEST=ECC_BI_03_I, ECC_BI_04_I, ECS_BV_03_I, ECS_BV_02_I

Change-Id: I5052b0d3eb0a41945a0c3919013458d90d894e6e
Reviewed-on: https://chromium-review.googlesource.com/206993
Tested-by: MengHuan Yu <menghuan@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>