CRAS: audio_thread_log - Make log entries fixed length.
Make audio log entries all the same length.  This will make it easy to
find the beginning of the log when printing it in the test client in a
later change, and is faster than trying to make the start of the tag
with a special sequence and escaping other data that matches.

BUG=none
TEST=none

Change-Id: I542733b8ff23a040ca55917a0e8433f1deccf306
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/207737