CRAS: mix - Fix bug in cras_mix_add_stream
When there is zero sample queued in shm, also update the number of
rendering samples to zero.
This fixes a potential bug in the write audio path: when playing back
to two streams of different buffer size, we could incorrectly update
the read offset in shm when there's actually no buffer to read.

BUG=None
TEST=None

Change-Id: I539927bb73f3cd466d7277da1f0c61075353cab8
Reviewed-on: https://chromium-review.googlesource.com/211084
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>