CRAS: audio_thread - Don't mix stream with skip_mix = 1
When two output streams are added and one is held by a bad client
that never fills audio to shm, we should skip mixing it so that
the other streams won't be affected.

Note that this doesn't solve the bad client problem completely, but
makes the other streams have a much better quality.

BUG=chromium:404885
TEST=Play youtube and at the same time execute 'cras_test_client
--playback_file -'

Change-Id: Id3dc4c87f3c68c283406f69bf76f7350849a4f22
Reviewed-on: https://chromium-review.googlesource.com/213045
Reviewed-by: Chinyue Chen <chinyue@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>