CRAS: audio_thread - remove fd from io stream
The fd parameter was just used as a cache.  Remove that and instead
get it from the stream when needed.  This is done so that
cras_io_stream can have per-stream details removed.  Eventually
cras_io_stream will be used for a stream to device mapping, and the
same stream could be held by multiple devices.

BUG=none
TEST=unittest

Change-Id: I9f82d3af07c6f29270d3273105213cd652b1ce10
Reviewed-on: https://chromium-review.googlesource.com/217154
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>