CRAS: move stream mixing to dev_stream.
Move the shm interaction and some decision making to dev_stream as
this is a property of the stream.  cras_mix no longer needs to know
about shm and its unit test is now much cleaner.  This paves the way
to add sample rate conversion logic to dev_stream in a later commit.

BUG=none
TEST=make check, play a few simultaneous streams.

Change-Id: I5fbb3c670e86393a0c8d3454e29c4bf587a5015c
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/217157
Reviewed-by: Hsinyu Chao <hychao@chromium.org>