CRAS: audio_thread - Handle dev error during drain.
If get_buffer returns an error, propagate that back up so that the
device is closed immediately.  If this error is ignored during
draining, it will stick the audio thread in a tight loop constantly
retrying.  Instead of that, since the device is already draining, if
an error happens abort draining and close the device.

BUG=chrome-os-partner:31818
TEST=play several youtube videos.

Change-Id: I6f2112839f55af3a51574283b5efdbd585bb5549
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/216701
Reviewed-by: Hsinyu Chao <hychao@chromium.org>