CRAS: Add --version parameter to cras_test_client
In the context of crouton, it is useful to know the exact commit
id that was used to build cras. And it can be useful for
development purpose as well.

If -DVCIS= is included in CPPFLAGS (Chromium OS build), use
that as version.
If not, generate a header file that defines the symbol from the
output of `git rev-parse`.

BUG=chromium:398796
CQ-DEPEND=CL:217540
TEST=cras_test_client --version outputs the correct git commit id.

Change-Id: I6049cb1e180c3a6cfe43bc33adbd12bd88f6f738
Reviewed-on: https://chromium-review.googlesource.com/217523
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>