CRAS: iodev - Set default channel count/layout
When an iodev fails to update channel layout for the format
of the appending stream is using. Fall back to use a default
channel count and matched layout, and let format conversion
do the rest of the work.

This is to fix a bug on ARM and BYT devices, and any USB
headsets don't support configuring channel map. The channel_layout
of the format could be not matched with its channel count, hence
causing record issues.

BUG=chromium:415412
TEST=unittest, youtube.com/my_webcam on Pixel and diasy

Change-Id: Ie99ed6cbd9ae9ccb2ade1b7e8a7f26cc292c6bd3
Reviewed-on: https://chromium-review.googlesource.com/218802
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>