CRAS: Fake select_device for Chrome UI
The Chrome UI needs select_device to work, so implement the behavior
as a combination of removing all active devs and adding the new dev to
the active list.  This will maintain the same behavior that users
expect.  The select_node api will later be removed once all users are
converted.

Change-Id: Ib1e513978bc9531af43a5aca59aac212c529110e
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/219980