CRAS: alsa_jack - Create gpio jacks on first input/output device
We used to create gpio jacks only on device 0 of a card. This does
not work for Samus since there are two front end DAIs:
device 0: System Playback (CRAS_STREAM_OUTPUT)
device 1: Capture (CRAS_STREAM_INPUT)

With this patch, headphone/HDMI gpio jacks are created and attached
to the first output device of a card. Mic gpio jacks are created and
attached to the first input device of a card.

BUG=chrome-os-partner:29649
TEST=On Samus bdw-rt5677 Mic Jack appears with this patch

Change-Id: I87ad98c7605a7855af86bfaaec8425c02a89427d
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/220591