CRAS: audio_thread - Handle odev errors
Handle errors on output devices.  If an output device returns an
error, remove it from the active list.  Streams that were playing on
that device will be moved to the empty device if no other device is
active.  In do_playback, don't cache the pointer to the first active
device as it can change if there is an error.

BUG=chromium:419945
TEST=play and record audio while inserting and removing devices.

Change-Id: Iff0983e6ff98e9dd8e19f19d36e02e8344ee9af6
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/221364
Reviewed-by: Hsinyu Chao <hychao@chromium.org>