CRAS: linear_resampler - increase resample precision
When doing SRC, the actual rate is trimmed to the floor
integer before use and causing a precision lost. Multiply
the numerator and denominator of the rate factor by 100,
so it takes more than 1 day for the lost frames count to be
accumulated to 1024.

BUG=chromium:420589
TEST=Play youtube to two active devices

Change-Id: I6bcac51e01ea6deba16462b201c783ea1a85f96e
Reviewed-on: https://chromium-review.googlesource.com/221706
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>