CRAS: iodev - Get estimate rate ratio
Change the getter to return the ratio of estimated rate and
the rate device is claimed to use. In later changes when we
align an active device to the rate of the master active device.
Using this ratio will make the code simpler as opposed to using
the estimate rate value directly.

BUG=chromium:420589
TEST=None

Change-Id: Ib6109ed86c06379dcfb8371d1189b3f28a78315f
Reviewed-on: https://chromium-review.googlesource.com/224624
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>