CRAS: fmt_conv - Add a default channel copy function
Adds a default channel conversion that copies a little of each
channel to all of the output channels.  It divides the input signal by
the number of channels and adds each input channel.  All output
channels play the same thing.  This is slow, and attenuates the signal
heavily, but it is better than dropping it.  It will only be used when
no better converter is found.

BUG=chromium:432425
TEST=play 8 channel test file from iis.fraunhofer.de

Change-Id: I6ada1281ae8b957f3a7d66d4c827d4eae9542e99
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/229730