UCM config: Samus: add auto channel swap on headphone plug/unplug
The speaker amps expect left channel data on rising edges of the
PDM clock, but the codec outputs left channel data on the falling
edges by default. To workaround the issue, this patch keeps the
channels flipped when there is no headphone, and swaps the channels
back to normal when headphone is plugged in.

BUG=chrome-os-partner:32809
TEST=left and right channel testing on headphones and speakers

Change-Id: I5fc6d6876563afb78712a8795f75cc0fa0f505fd
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/231077
Reviewed-by: Anatol Pomazau <anatol@google.com>