CRAS: bt_profile - Use bt device to replace transport
The transport object carries information and is used in Bluez's
media API. We borrowed it for the cras_bt_profile module and
it turns out redundant in cases when we want to associate a profile
or HFP related objects with cras_bt_device.
This change refactors the code to better align with Bluez profile
API and will allow the device disconnected event be notified to bt
profiles in a clean way.

BUG=chromium:432434
TEST=Test A2DP playback. Manual enable HFP, test playback & capture.

Change-Id: I2fab0b32965227512ad841bd7df785cd1af4a2ab
Reviewed-on: https://chromium-review.googlesource.com/231864
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>