CRAS: bt_io - API to switch between A2DP and HFP/HSP
At events that a bluetooth device opens or closes, we might
want to switch to a different profile. When switches the profile
of a bt device, it is required to reinitialize the bt_ios.
This change adds the function to temporarily close the active
bt_ios and activate them later so they can switch to the preferred
profile.

BUG=chromium:432434
TEST=unittest

Change-Id: I7ac2feab38b1dd4eb4a4def0bbc2323eb5842bbd
Reviewed-on: https://chromium-review.googlesource.com/231870
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>