CRAS: bt_io - Add bt_io module
To allow A2DP and HFP/HSP iodev to exist at the same time and can
switch between them. Adds the bt_io module as a wrapper to hold a
list of profile specific iodevs and delegate the audio sample
transmission job to the iodev of active profile.
In later change the bt_io will detect the situation user wants to
switch to a different proflie and communicate with cras_bt_device
to do the switch.

BUG=chromium:432434
TEST=unittest

Change-Id: I4dcd7269b27f012115676cb44b9b60cb0f9cda6d
Reviewed-on: https://chromium-review.googlesource.com/231867
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>