CRAS: bt_device - try remove profile dev
There could be a case that BT profiles got disconnected one at
a time so there's a short period audio_thread is still accessing
the device before the bt_io destructor is called.  For safety we
need to check if the profile iodev is open for use before we try
to remove it.

BUG=chromium:432432
TEST=Connect bluetooth headset, play YouTube, disconnect headset.

Change-Id: I1d5bf3a5e77fcec2c4af0727b985bd2f567904c2
Reviewed-on: https://chromium-review.googlesource.com/234623
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>