CRAS: bt_io - Don't remove the dummy node
The first ionode created on a bt_io is a dummy node used to
point to the other node of active profile (A2DP or HFP). Be
careful not to remove this node when we have to need to remove
any other node when BT device disconnects.

BUG=chromium:446089
TEST=Connect BT headset and test HFP using audio visualizer app.

Change-Id: Iebd3fa59371cfc8f907d49b57b7251460461d2ff
Reviewed-on: https://chromium-review.googlesource.com/238319
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>