ucm: ryu: adjust default ADC1 capture volume after TLV change
When we fixed the volume TLV, the resolution of
"ADC1 Capture Volume" was halved, but its default
value in the ucm config was not changed:

UPSTREAM: ASoC: rt5677: make volume TLV closer to reality
https://chromium-review.googlesource.com/233550

A value of 63 used to mean -17.625+63*0.375 = 6dB; but after
the patch, it means -17.625+(63*2+1)*0.375 = 30dB now. With
a "STO1 ADC Boost Volume" of 24dB, the overall gain is too
big and causing distortion.

This patch changes the default value to 31 which means
-17.625+(31*2+1)*0.375 = 6dB.

Based on Samus change https://chromium-review.googlesource.com/#/c/242741/

BUG=None
TEST=P3 mic recording

Change-Id: Ida7619e6c2972529549c1d9c75e5cf26511d7209
Reviewed-on: https://chromium-review.googlesource.com/247231
Reviewed-by: Ben Zhang <benzh@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Anatol Pomazau <anatol@google.com>
Commit-Queue: Anatol Pomazau <anatol@google.com>
Trybot-Ready: Anatol Pomazau <anatol@google.com>