CRAS: HFP - Fix response string for indicator test command
An established service level connection is required for HFP
headset, and the indicator response is used during the SLC
initialize process. The original bad formated string is a
suspicious cause to incomplete connection for some BT headsets
when powers off and back on and then initiate a automatic
re-connection.

BUG=chromium:449068
TEST=Test 'Avantree Hive' headset, while it connects and streaming
audio to Chromebook, power it off to make audio switch to internal
speaker. And then power it back on, wait for it to reconnect, see
audio transfer back to headset.

Change-Id: Icf5843973f4d39e0065bcdfa6b08c18cf94f74c6
Reviewed-on: https://chromium-review.googlesource.com/243528
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>