CRAS: SLC - Fix crash when no init_cb for SLC establishment
When headset is connected in HSP mode, it is not required to
wait for service level connection being established before setting
up any audio connection.
Fix the crash when trying to call the NULL init_cb in this
case. And also modify the handling logic to consider SLC is
fully initialized after successfully responded with OK regardless
of the indicator update status, which is more aligned with spec.

BUG=chromium:463058
TEST=Manual test HFP/HSP only headset, and A2DP + HFP/HSP headset

Change-Id: Id60a64b3cfac0084d7cd5f2466e3cd9ed69709ee
Reviewed-on: https://chromium-review.googlesource.com/255032
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>