CRAS: hfp_ag - Handle repeat HFP/HSP connection
In rare cases using HFP/HSP only headset, CRAS receives two
consecutive connection event for the same device, one for HFP
and the other for HSP. Since we don't have a preference to use
HFP or HSP for one headset, CRAS handles this scenario by tear
down what was setup for the 1st connection and proceed with the
2nd connection.
However we're seeing issues that audio channel failed to open when
headset connects, with error "Device or resource busy (16)".  And
the related shows that this audio failure happens after the repeat
HFP/HSP connection in a short time.
This change attempts to address this issue by ignore the connection
event if an audio gateway has already been setup for the same device.

BUG=chromium:448594, chromium:457015
TEST=Connect HFP/HSP only headset and play audio.

Change-Id: I4d615388d4520a934ee747b35b640596b5eb5e67
Reviewed-on: https://chromium-review.googlesource.com/259582
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>