CRAS: audio_thread - Remove unused metrics
The logging of the longest timeout has been disabled for a while, and no
one seems to be missing it.  Remove the residue of the feature.  It can
be added back in if needed later.  After the coming changes to keep a
stream list in the main thread it will be easier anyways.

Change-Id: Ie9c341961ffc8be4ffbe9c350fe4f5a40c1ad6ed
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/261324