CRAS: Open iodevs from the main thread.
This is the bulk of the changes needed to move open and close calls to
iodevs out of the audio thread.  Having them in the audio thread is
causing glitching because they take an undetermined amount of time.
Commits following this one will address draining of streams and devices.
With this commit stream and device draining are broken.

iodev_list now manages which devices are opened. This is determined by
when a stream is added to the device.  rclient adds streams to the
system, iodev_list registers a callback for when a stream is added or
removed.  iodev_list then determines which devices should play the
stream, opens them if necessary and tells the audio thread the device is
open and which streams the device should service.

Change-Id: I8c462f6ef3075fc01d7fe3f65c53cd67fdcbc260
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/262510
Reviewed-by: Chinyue Chen <chinyue@chromium.org>
Commit-Queue: Chinyue Chen <chinyue@chromium.org>
Tested-by: Chinyue Chen <chinyue@chromium.org>