CRAS: Drain streams before removing
Before deleting a stream, ensure that all the samples in shared memory
have been sent to the device.  This logic now resides in the main
thread, it used to be in the audio_thread.  Having it in the main thread
helps simplify the audio thread which was getting complicated and it
allows for stream lifespan to be totally managed in stream_list.

Change-Id: Id826f6241edb5d5d30a2c7be53f987417a7218c7
Reviewed-on: https://chromium-review.googlesource.com/262766
Reviewed-by: Chinyue Chen <chinyue@chromium.org>
Commit-Queue: Chinyue Chen <chinyue@chromium.org>
Tested-by: Chinyue Chen <chinyue@chromium.org>