CRAS: audio_thread - fill odevs without streams
If an odev is running and there aren't any streams attached fill them
with zeros.  This will allow for the device to be kept open after the
streams are removed, allowing for it to drain completely and resume
faster if a new stream is added.

Change-Id: I807535c12260458291c008bf9cdbe982720d7750
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/262962
Reviewed-by: Chinyue Chen <chinyue@chromium.org>
Commit-Queue: Chinyue Chen <chinyue@chromium.org>
Tested-by: Chinyue Chen <chinyue@chromium.org>