CRAS: audio_thread - Wake up sooner when draining a dev
When the last stream is removed from an output device, the device has
zeros played to it for a while before closing.  It needs to be filled
with silence before it underruns.  Try to keep the device filled to the
min_cb_level by only sleeping until that many samples remain.

BUG=chromium:471506
TEST=play youtube, close window, no audio disturbance.

Change-Id: Ic4ca531d7667294d970dfb25f9474e48cbede03a
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/264787