CRAS: iodev - Correctly update min/max cb level
We need to pre set min/max cb level to zero and buffer size
in remove stream function. Or the later update could go wrong
when iterate through all streams, if the stream to remove was
the one of min/max cb threshold of all streams. Also keep the
last min_cb_level on iodev when the last stream is removed, this
level will be used for iodev draining.

BUG=chrome-os-partner:38999
TEST=None

Change-Id: I2ec124e473313f548251a68fe66a78b2e7e47fc9
Reviewed-on: https://chromium-review.googlesource.com/265391
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>