CRAS: audio_thread - Only wait as long as needed for a skipped callback.
When a playback callback has to be skipped because there isn't an open
buffer, try again as soon as there is space.  Before an entire period
would be skipped.  Now, ignore wake ups from that device until there is
a buffer available, and make the callback as soon as there is space.

BUG=chrome-os-partner:38183
TEST=play google play music while screen sharing in a hangout, listen
for dropouts.

Change-Id: I0d8e064f57521b22da06a28585055673427c6f9f
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/266768