CRAS: iodev_list - Fix is_enabled flag
Rename the is_active flag to is_enabled. This flag was set/unset
in audio thread when an iodev opened/closed, which breaks the
original purpose of the flag. Move it to be set/unset in main
thread when the iodev gets enabled/disabled.

BUG=chromium:482233
TEST=cras_test_client --dump_s

Change-Id: I2c20191741e9a9abb79216025da7a62cc3c17d0d
Reviewed-on: https://chromium-review.googlesource.com/271716
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Chinyue Chen <chinyue@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>