CRAS: linear_resampler - Fix resample frames calculation
Originally the in/out frames calculation is not accurate enough,
that it always result a number sightly larger then the actual
number of converted frames, which could cause problem in the
usage of chained SRC.
Add into account the offset state of the linear resampler, so
the in <-> out frames calculation aligns with the exact converted
frame numbers of the next call to linear_resampler_resample().

BUG=chromium:460515
TEST=Apply whole patch set, open flash video(44.1KHz, 2048 fr)
and a web audio demo page(48KHz, 1024 fr) with two Jabra
speakerphones enabled. Verify audio does not stop when play
for an hour long.

Change-Id: I144854a5afca7c73c1efff8af9d49ff0fab4ace2
Reviewed-on: https://chromium-review.googlesource.com/276839
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>