CRAS: linear_resampler - Fix corner cases
In linear_resampler_resample function, if dst_frames is zero
then it's incorrect to iterate dst_idx from zero and it will
leak one frames in the source buffer.
If src_frames points to an unsigned of value 0, we should not
minus one from this unsigned value and then use it to bound
index.

BUG=chromium:460515
TEST=unittest

Change-Id: I9d7aff853666532701d528194e2ac274a3f0c6af
Reviewed-on: https://chromium-review.googlesource.com/277528
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>