CRAS: audio_thread - Don't free format if add stream fail
This error handling call was added to prevent crash back when we
manage iodev open/close in audio thread. This code makes no sense
now because iodev is added to audio thread only when opened, and
it's allowed to run with no stream attached.

BUG=None
TEST=None

Change-Id: I45542ae9a136065cc3a989f041fce2b03f738558
Reviewed-on: https://chromium-review.googlesource.com/280346
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>