CRAS: dev_stream - Add function to check if stream can fetch
Add function dev_stream_can_fetch() and use it to determine if it's
okay to request playback data for a stream. This new function is
used to make the code more readible in output fetch and sleep
scheduling.

BUG=None
TEST=Manual test normal playback and use cras_test_client using 'p'
key to pause/resume the client process and verify audio turn silent
and recover smoothly.

Change-Id: I83ac7e95d4a17698575ce93e0edead78439d7597
Reviewed-on: https://chromium-review.googlesource.com/289421
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>