CRAS: server_metrics - Log longest fetch delay
Each output stream is fetched regularly base on the stream's
callback threshold. If a client fails to respond to a fetch
request in time, the next fetch request will be delayed.

This change adds to cras_server_metrics the ability to log
various types of information to UMA. And use it in audio thread
to track the longest case of stream's fetch delay.

BUG=chromium:335774
TEST=cras_test_client --playback_file and use 'p' to suspend
the client and the resume it, repeat this many times then
visit chrome://histogram to verify the Cras.StreamTimeoutMilliSeconds
histogram changed accordingly.

Change-Id: I81862cb2cbd2e62f78f1dcef13dafc30c34fee65
Reviewed-on: https://chromium-review.googlesource.com/289429
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>