CRAS: audio_thread - Add stream to only enabled devices
When adding a non-pinned stream, this patch ensures the stream is
added to only enabled devices instead of all devices. This avoids
the issue where a normal capture stream (e.g. from hangouts) gets
added to the AOKR (hotwording) device which is openned due to a
pinned stream but not enabled. The AOKR device won't have any samples
until a hotword is detected, so the normal capture stream is blocked.

BUG=chrome-os-partner:42983
TEST=Enable hotwording on samus, start hangouts, listen from the
remote side and verify mic works.

Change-Id: I0ceb05be0e360ef7a62b5ab707d78ddd975ef1be
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/289965
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>