CRAS: alsa_mixer - Check input control when creating alsa mixer.
For usb devices, the names of their capture controls may be 'Mic'
or 'Microphone' or something else. We need to identify them and
associate to a input node so that we can mute them properly.

BUG=chromium:526092
TEST=Plug in Plantronic .Audio 655 DSP.
Use 'cras_test_client --capture_mute 1' to mute the Mic and verify
the result with 'amixer -c <CARD>'.

Change-Id: Ib6d6e67b0b1168d3f6945aa949e6b685e9b09076
Reviewed-on: https://chromium-review.googlesource.com/296876
Commit-Ready: Yao-Wen Mao <yaowen@google.com>
Tested-by: Yao-Wen Mao <yaowen@google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>