CRAS: alsa_jack - Checks device name for GPIO jack
Checks if a GPIO jack should be associated with a device by checking the
device name if it is specified in UCM. Otherwise, assume GPIO jacks
should be associated with the first device.
On the other hand, hctl jacks still use pcm=? in its name to determine the
device index because it is the common case on hctl jacks.

BUG=chrome-os-partner:40529
TEST=make check
TEST=on samus, test headphone and microphone.
TEST=on oak, HDMI jack works.

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: I4e3397dd95cbac290f6daa61b5029d3068cfd64b
Reviewed-on: https://chromium-review.googlesource.com/309094