CRAS: system_state - use posix shm
Use posix shm instead of SysV shm.  This is done so that the client can
be passed an open fd to the shm region.  Passing the fd is more secure
since before anyone in the cras group could access the state.  More
importantly this allows for the client to access the state even when it
is running in a differen IPC namespace than the server.

Change-Id: Iedb1fc753dc0fdc415862cb96ca97ceae3a8111d
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/312771
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Chinyue Chen <chinyue@chromium.org>