CRAS: iodev - Use rate in format when allocating DSP
DSP should use the actual sampling rate of device because it is applied
right before writing samples to device.
Currently ext_format and format only differ in num_channel, but they
have different meaning so we should make this logically correct in case
we allow changes in sampling rate in ext_format in the future.

BUG=chrome-os-partner:47898
TEST=Plays a 44100 rate stream. Checks the rate of DSP is 48000 on
samus.

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: Ib9e1d486f8eb48020c89db670a27d7e9eedd11f7
Reviewed-on: https://chromium-review.googlesource.com/313852
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>