UCM: daisy_spring: Let Cras Handle output switch
Speaker, Headphone and HDMI are on the same device. When user
switches from headphone to HDMI, speaker will get incorrectly
turned on because it is in Headphone device disable sequence.
Remove the output control in disable sequence and enable sequence in UCM and
let Cras handle the switch through simple mixer interface.

BUG=chromium:560541
TEST=switch from headphone to HDMI, check speaker switch is off.
TEST=switch between headphone, speaker, and HDMI, make sure sound goes
to the right node.
Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>

Change-Id: I614aaa37ccbcf7d883189a8c208e39de2d5f21fc
Reviewed-on: https://chromium-review.googlesource.com/313909
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Chinyue Chen <chinyue@chromium.org>