UCM: daisy_skate: Let Cras handle output switch
Speaker, Headphone and HDMI are on the same device. When user
switches from headphone to HDMI, speaker will get incorrectly
turned on because it is in Headphone device disable sequence.
Remove the output control in disable sequence and enable sequence in UCM
and let Cras handle the switch through simple mixer interface.

BUG=chromium:560541
TEST=switch from headphone to HDMI, check speaker switch is off.
TEST=switch between headphone, speaker, and HDMI, make sure sound goes
to the right node.
Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: Ib50f2b373a113d94c0716c2b1b3e4c6916acc6e0
Reviewed-on: https://chromium-review.googlesource.com/314418
Reviewed-by: Chinyue Chen <chinyue@chromium.org>