CRAS: iodev_list: Enable fallback device before disabling the only device
In cras_iodev_list_disable_dev, if the device to be disabled is the only
enabled device, we should enable fallback device first.
This call happens when a node is unplugged.
Without this patch, the fallback device will be added some time
later after original device is removed. Although this is a small time
window, we should enable fallback device first to minimize the time
window where there is no device to play data. On Samus this time window
is about 5 ms, so this is not actually noticeable to users.

BUG=chrome-os-partner:49101
TEST=plug/unplug 3.5mm jack, check audio thread log, see device 2
(fallback device) is enabled before device 8 is disabled.
$ grep -E 'DEV_ADDED|DEV_REMOVED' log
     70091.549208479  DEV_ADDED                      dev:2
     70091.549339569  DEV_REMOVED                    dev:8
     70091.675618425  DEV_ADDED                      dev:8
     70091.675928957  DEV_REMOVED                    dev:2

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: I03ec9fecdd4d857d2552584f424cb707131378ba
Reviewed-on: https://chromium-review.googlesource.com/320672